# shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cab Dispatch API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tariffs": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "List Tariffs",
        "operationId": "list_tariffs_v1_tariffs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Region"
            }
          },
          {
            "name": "booked",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Booked"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TariffRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tariffs"
        ],
        "summary": "Create Tariff",
        "operationId": "create_tariff_v1_tariffs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TariffCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tariffs/active": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "Get Active Tariff",
        "description": "Resolves the tenant's currently-effective tariff for `region` at `at`\n(defaults to now). NOTE: registered before `/{tariff_id}` so \"active\"\nis never captured as a path parameter.",
        "operationId": "get_active_tariff_v1_tariffs_active_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Region"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO8601 instant; defaults to now",
              "title": "At"
            },
            "description": "ISO8601 instant; defaults to now"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tariffs/{tariff_id}": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "Get Tariff",
        "operationId": "get_tariff_v1_tariffs__tariff_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tariffs"
        ],
        "summary": "Update Tariff",
        "operationId": "update_tariff_v1_tariffs__tariff_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TariffUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tariffs"
        ],
        "summary": "Delete Tariff",
        "operationId": "delete_tariff_v1_tariffs__tariff_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tariffs/{tariff_id}/extras": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "List Extras",
        "operationId": "list_extras_v1_tariffs__tariff_id__extras_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExtraRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "tariffs"
        ],
        "summary": "Create Extra",
        "operationId": "create_extra_v1_tariffs__tariff_id__extras_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtraCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tariffs/{tariff_id}/extras/{extra_id}": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "Get Extra",
        "operationId": "get_extra_v1_tariffs__tariff_id__extras__extra_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          },
          {
            "name": "extra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Extra Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "tariffs"
        ],
        "summary": "Update Extra",
        "operationId": "update_extra_v1_tariffs__tariff_id__extras__extra_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          },
          {
            "name": "extra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Extra Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtraUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtraRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tariffs"
        ],
        "summary": "Delete Extra",
        "operationId": "delete_extra_v1_tariffs__tariff_id__extras__extra_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          },
          {
            "name": "extra_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Extra Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tariffs/{tariff_id}/change-log": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "List Tariff Change Log",
        "operationId": "list_tariff_change_log_v1_tariffs__tariff_id__change_log_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_TariffChangeLogRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tariffs/{tariff_id}/change-log/{log_id}": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "Get Tariff Change Log Entry",
        "operationId": "get_tariff_change_log_entry_v1_tariffs__tariff_id__change_log__log_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tariff_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tariff Id"
            }
          },
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffChangeLogRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/fares-order/current": {
      "get": {
        "tags": [
          "tariffs"
        ],
        "summary": "Get Current Fares Order",
        "description": "Returns the platform-wide (tenant_id IS NULL) Fares Order reference\ntariff currently effective for `region`. 404s until the integration step\nseeds the global reference rows \u2014 expected in isolation.",
        "operationId": "get_current_fares_order_v1_fares_order_current_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "urban",
              "title": "Region"
            }
          },
          {
            "name": "at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO8601 instant; defaults to now",
              "title": "At"
            },
            "description": "ISO8601 instant; defaults to now"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TariffRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ExtraRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExtraRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[ExtraRead]"
      },
...
      "Page_TariffRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TariffRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "skip": {
            "type": "integer",
            "title": "Skip"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "skip",
          "limit"
        ],
        "title": "Page[TariffRead]"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}